FEAT:[Modal][Vadim]: added modals with animations
1) FADE IN
2) SLIDE IN DOWN
3) FADEIN UP
4) SLIDE IN UP
5) ZOOM IN UP
