+++ package.json
@@ -10,24 +10,25 @@
 		"preview": "vite preview"
 	},
 	"dependencies": {
+		"@headlessui/react": "^2.1.8",
 		"@tanstack/react-query": "^5.56.2",
 		"axios": "^1.7.7",
 		"class-variance-authority": "^0.7.0",
 		"clsx": "^2.1.1",
-		"lucide-react": "^0.441.0",
+		"lucide-react": "^0.446.0",
 		"react": "^18.3.1",
-		"react-datepicker": "^7.3.0",
+		"react-datepicker": "^7.4.0",
 		"react-dom": "^18.3.1",
 		"react-hook-form": "^7.53.0",
 		"react-select": "^5.8.1",
-		"sass": "^1.79.2",
-		"sweetalert2": "^11.14.0",
+		"sass": "^1.79.3",
+		"sweetalert2": "^11.14.1",
 		"tailwind-merge": "^2.5.2",
 		"tailwindcss-animate": "^1.0.7"
 	},
 	"devDependencies": {
 		"@eslint/js": "^9.10.0",
-		"@tanstack/eslint-plugin-query": "^5.56.1",
+		"@tanstack/eslint-plugin-query": "^5.58.1",
 		"@trivago/prettier-plugin-sort-imports": "^4.3.0",
 		"@types/node": "^22.5.5",
 		"@types/react": "^18.3.8",
@@ -41,11 +42,11 @@
 		"globals": "^15.9.0",
 		"postcss": "^8.4.47",
 		"prettier": "^3.3.3",
-		"prettier-plugin-tailwindcss": "^0.6.6",
-		"tailwindcss": "^3.4.12",
+		"prettier-plugin-tailwindcss": "^0.6.8",
+		"tailwindcss": "^3.4.13",
 		"typescript": "^5.6.2",
-		"typescript-eslint": "^8.6.0",
-		"vite": "^5.4.6"
+		"typescript-eslint": "^8.7.0",
+		"vite": "^5.4.8"
 	},
 	"packageManager": "pnpm@9.10.0+sha512.73a29afa36a0d092ece5271de5177ecbf8318d454ecd701343131b8ebc0c1a91c487da46ab77c8e596d6acf1461e3594ced4becedf8921b074fbd8653ed7051c"
 }
