+++ src/index.scss
@@ -1,4 +1,4 @@
-// react-select library custom styles
+@import url('./assets/css/animate.css');
 
 @tailwind base;
 @tailwind components;
